infra(docs): migrate docs deploy to cloudflare pages
