+++ package.json
@@ -30,7 +30,7 @@
     "docs:dev": "pnpm --filter robota-docs run dev",
     "docs:build": "node scripts/docs/prepare-docs.js",
     "docs:preview": "pnpm --filter robota-docs run preview",
-    "docs:deploy": "pnpm docs:build && node scripts/docs/deploy-gh-pages.mjs",
+    "docs:deploy": "pnpm docs:build && node scripts/docs/deploy-cloudflare-pages.mjs",
     "docs:validate-structure": "node scripts/docs/validate-package-docs-structure.mjs",
     "dag:dev": "pnpm --filter @robota-sdk/dag-designer --filter @robota-sdk/dag-runtime-server --filter @robota-sdk/dag-orchestrator-server --filter @robota-sdk/dag-studio -r --parallel dev",
     "dag:build": "pnpm --filter @robota-sdk/dag-core build && pnpm --filter @robota-sdk/dag-api build && pnpm --filter @robota-sdk/dag-orchestrator build && pnpm --filter @robota-sdk/dag-designer build",
